# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a high-performance dating app backend server.",
        "title": "Dating App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/profile": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes the authenticated user's account and all associated data.",
                "tags": [
                    "users"
                ],
                "summary": "Delete user account",
                "responses": {
                    "200": {
                        "description": "Successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's profile information, including photos and preferences.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Profile update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated profile",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/profile/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the profile details of a specific user by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/users/upload-url": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a temporary S3 upload URL for private user media.",
                "tags": [
                    "users"
                ],
                "summary": "Get private upload URL",
                "responses": {
                    "200": {
                        "description": "Upload URL details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UploadURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/upload-url/public": {
            "get": {
                "description": "Generates a temporary S3 upload URL for publicly accessible assets (e.g., registration photos).",
                "tags": [
                    "users"
                ],
                "summary": "Get public upload URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client Identifier",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload URL details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UploadURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/verify-face": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Performs facial verification by comparing an uploaded snapshot against the user's profile photo.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Verify user face",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Face snapshot",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.VerificationResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.UserStatus": {
            "type": "string",
            "enum": [
                "onboarding",
                "active",
                "banned"
            ],
            "x-enum-varnames": [
                "UserStatusOnboarding",
                "UserStatusActive",
                "UserStatusBanned"
            ]
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "response.ConsumableItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "item_type": {
                    "type": "string"
                }
            }
        },
...
        "response.MasterItemResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.PhotoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "consumables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ConsumableItemResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "height_cm": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "interested_genders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "location_city": {
                    "type": "string"
                },
                "location_country": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "main_photo": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PhotoResponse"
                    }
                },
                "relationship_type": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "status": {
                    "$ref": "#/definitions/entities.UserStatus"
                },
                "subscription": {
                    "$ref": "#/definitions/response.UserSubscriptionResponse"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
        "response.UserSubscriptionResponse": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "plan_id": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
        "response.VerificationResult": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "error": {
                    "type": "string"
                },
                "is_match": {
                    "type": "boolean"
                }
            }
        },
...
        "user.PhotoRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "_destroy": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "user.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "date_of_birth": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "height_cm": {
                    "type": "integer"
                },
                "interested_in": {
                    "type": "string"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "location_city": {
                    "type": "string"
                },
                "location_country": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "photos": {
                    "description": "List of photo objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.PhotoRequest"
                    }
                },
                "relationship_type": {
                    "type": "string"
                },
                "status": {
                    "description": "onboarding, active",
                    "type": "string"
                }
            }
        },
        "user.UploadURLResponse": {
            "type": "object",
            "properties": {
                "file_key": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}